requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
distro==1.9.0; python_version >= '3.6'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.8.2; python_version >= '3.8'
openai==1.65.3; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pysimplegui==5.0.99.99; python_version >= '3.6'
python-dotenv==1.0.1; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
sniffio==1.3.1; python_version >= '3.7'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
